+++ Source/Plugins/SurfaceMeshing/SurfaceMeshingFilters/LaplacianSmoothing.h
@@ -187,7 +187,7 @@ class LaplacianSmoothing : public SurfaceMeshFilter
 
   private:
     DEFINE_REQUIRED_DATAARRAY_VARIABLE(int8_t, SurfaceMeshNodeType)
-    DEFINE_REQUIRED_DATAARRAY_VARIABLE(int8_t, SurfaceMeshFaceLabels)
+    DEFINE_REQUIRED_DATAARRAY_VARIABLE(int32_t, SurfaceMeshFaceLabels)
 
 #if OUTPUT_DEBUG_VTK_FILES
     void writeVTKFile(const QString& outputVtkFile);
